pyproject.toml
[project]
name = "pendulum-tools"
version = "0.2.0"
description = "Post-processing and upload tools for double pendulum simulations"
readme = "README.md"
authors = [
    { name = "Jérémie Martin", email = "martin.jeremie42@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pysubs2>=1.7.0",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pyqt6>=6.10.1",
]

[project.scripts]
pendulum-tools = "pendulum_tools.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pendulum_tools"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
